Remove Grok for subscribers
